Modify reconstruction_dilation_step dtype

+++ layers.py
@@ -239,7 +239,7 @@ def reconstruction_dilation_step(X):
     # perform a geodesic dilation with X[0] as marker, and X[1] as mask
     X[0]=tf.keras.layers.MaxPooling2D(pool_size=(3, 3),strides=(1,1),padding='same')(X[0])
     X[0]=tf.keras.layers.Minimum()([X[0],X[1]])
-    return tf.cast(X[0], tf.double)
+    return X[0]
 
 def reconstruction_dilation(X,steps=None):
     """
